wavesurfer only creates one audioContext, still need to test with multiple tracks
